Converted amount field in saved wealths to double from int because it would be more appropriate while calculations. (#16)
* Converted amount field of saved wealths to double

* duplicate codes refactored and moved into a util class

+++ wealth_calculator/lib/feature/calculator/view/calculator_view.dart
@@ -51,11 +51,11 @@ class CalculatorView extends StatelessWidget {
                     (wealth, amount) {
                       ItemDialogs.showEditItemDialog(
                         context,
-                        MapEntry(wealth, 0),
+                        MapEntry(wealth, 0.0),
                         (wealth, amount) {
-                          context
-                              .read<CalculatorBloc>()
-                              .add(AddOrUpdateCalculatorWealth(wealth, amount));
+                          context.read<CalculatorBloc>().add(
+                              AddOrUpdateCalculatorWealth(
+                                  wealth, amount.toDouble()));
                         },
                       );
                     },
